swimcu: Leave ULPM GPIOs configured during shutdown
GPIOs 36 and 38 must be left configured when the device is
entering low power mode. If the GPIOs are unconfigured before
shutting down the MDM, the MCU will never be able to detect
a wake up event.
This change also ensures the pull up resistors are enabled on
the GPIOs. The FX30 hardware design requires these resistors
to be enabled.

Resolves: LXSWIREF-1895 ALPC-1087
Relates: ALPC-879
Change-Id: Ic331c14766191ce2cda8a4bf3324c9d962817be2
Signed-off-by: Eric Hay <EHay@sierrawireless.com>
Signed-off-by: Zoran Markovic <zmarkovic@sierrawireless.com>