ssmem: Add support for Sierra Wireless shared memory
Content of arch/arm/mach-msm is moved to arch/arm/mach-qcom
in kernel 4.14.

Note, DM: There is still some MACH_MSM "residue" in kernel 4.14.
We should eliminate it to avoid confusion in the future. For now,
some components like lk depend on mach-msm name. In order to be
backwards compatible, we are adding mach-msm soft link:

        mach-msm -> mach-qcom

Resolves: LXSWIREF-1717
Change-Id: If6c6b3cf6388ab3859a3c9c5afe4d4ad792399e4
Signed-off-by: Dragan Marinkovic <dmarinkovi@sierrawireless.com>