gpio: Add SX150X gpio expander support
Enable the driver for Semtech SX150x-series I2C GPIO expanders used by MangOH
boards. Linux 4.10 onwards, the GPIO driver is replaced by a Pinctrl driver.
Adapt the DTS of MangOH boards accordingly.

Change in behaviour from 3.18: The gpioexpanders are no more statically
defined at gpiochip3nn; They are dynamically assigned a chip number
according to the order of initialisation.

Correction from 3.18: sierra_gpio.c:
 - When using the format "alias-NAME = <&device offset 0>" to define
an alias, if the device is not yet initialised the sierra_gpio driver
init is deferred to wait for the device init.
 - Ownership flag is updated anytime a define, export, unexport or show
is attempted on an aliased pin. Not just during the DTS parsing, as it
was in 3.18. This is to account for:
   1. Cases where the format "alias-NAME = <gpio-number 0>" is used to
      define an alias, but the chip is initialised after DTS parsing by
      sierra_gpio driver.
      The mask will not be available to be read at this time to
      correctly update the ownership flag. In 3.18, the ownership
      flag will be updated with "1" denoting apps. In 4.14, even the
      ownership flag will not be available during this case and hence will
      hold the default vlaue of "0" denoting modem. Both do not
      reflect the real status of ownership once the chip is initialized.

   2. When an alias is defined at run-time through /sys/class/gpio/.

Resolves: LXSWIREF-1720
Change-Id: Idb10f318ed35983c73be2814982441648deef096
Signed-off-by: Anju Kunjumon <akunjumon@sierrawireless.com>