sierra_gpio: Avoid repeated config after deferring
Take care of the after effects of deferring. Avoid re-scanning and
re-configuring of aliases and their sysfs entries in case of re-entry
from a deferred init.

Commit be4f3f defers sierra_gpio driver init, if an alias refers to a
gpiochip which is yet to be initialised. When the init is attempted again
after the gpiochip is initialised, do not restart processing of aliases
from the first in list; start from the deferred entry instead.

The return status is preserved accross entries.
Additionally, the debug levels of nominal info from driver, are reduced.

Resolves: LXSWIREF-1757
Change-Id: I9a681cd300a7237f7446ca85784a0a165d2371cb
Signed-off-by: Anju Kunjumon <akunjumon@sierrawireless.com>