mdm9x28: Fake existence of audio subsystem
Qualcomm audio packages assume existence of audio codecs in kernel.
Since codecs were not added to kernel as of yet, introduce headers
necessary for audio packages to build.

This is an interim patch until Qualcomm audio code is introduced
in kernel, at which point this patch could be reverted. There is
absolutely no intention, nor expectation for audio to work.

Resolves: LXSWIREF-1801
Change-Id: I36550d8180fd1f57c7736bae158d917fc1720d35
Signed-off-by: Zoran Markovic <zmarkovic@sierrawireless.com>