swimcu: implement MCU watchdog SYSFS interface
The MCU watchdog SYSFS interface is exposed if and only if this optioanl
feature is supported by the currently running MCUFW on the device. The
MCU watchdog is enabled with custom-defined timeout and reset delay via
SYSFS interface. On the watchdog timeout, MCU delivers a "watchdog"
event to MDM and expect MDM to renew the MCU watchdog before the reset
delay is elapsed.

On receipt of the "watchdog" event, the swimcu driver sends another
request to re-start the MCU watchdog. In addition, the swimcu driver
generates and delivers a MCU_WATCHDOG event to user space so that a
user-space application can use it to trigger additional operations
in sync with the device wakeup time for power saving.

MDM will be reset by MCU if a "watchdog" event is not received by MDM
upon watchdog expiry or the MCU watchdog timer is not restarted by MDM
before the reset delay is elapsed.

MCI protocol definition is updated for the new features.

Resolves: LXSWIREF-1747
Related: QTI9X07-581
Signed-off-by: Dylan Zhang <dyzhang@sierrawireless.com>
(from commit 59cbd8e32e1d234a7dfa58e84dd758733df9d5e6)
Signed-off-by: Ryan Kirk <rkirk@sierrawireless.com>
Signed-off-by: Dragan Marinkovic <dmarinkovi@sierrawireless.com>
Modified-by: Dragan Marinkovic <dmarinkovi@sierrawireless.com>
Change-Id: Icc86515fcecc53f31579efd50ace5fca0cca941b