swimcu: per-device calibration of MCU HW timer clock (1K LPO)
A mechanism is implemented for calibration of MCU 1K LPO, which drives
the MCU HW timers. An MCU HW timer is started with mcu_time immediately
when the I2C communication with MCU is established. A calibaration event
is delivered when the timer expires, indicating the end of calibration.
Corresponding mdm_time is calculated based on RAW MONOTOIC timestamps.

The "calibrate" SYSFS is exposed mainly for debug purpose, although it
can also be used to re-start a calibration if desired. Note that the
SWIMCU driver converts the user-input timeout value to corresponding
calibrated MCU time automatically. User need not and shall not convert
the inputs with the result in the "calibrate" SYSFS when they configure
the timeout value for MCU operations.

Resolves: LXSWIREF-1747
Related: QTI9X07-771
Signed-off-by: Dylan Zhang <dyzhang@sierrawireless.com>
Signed-off-by: Dragan Marinkovic <dmarinkovi@sierrawireless.com>
Modified-by: Dragan Marinkovic <dmarinkovi@sierrawireless.com>
(from commit ed6369706fb622af26354e7f3fdcd22107934c7d)
Change-Id: I8018bf651bef73c8e662b065b20323136194994e