swimcu: improve MCU LPO calibration procedure
This change resolves potential issues where ULPM is requested before
the MCU LPO calibration is completed. The request is always honored
after the calibration timer is stopped. New data are calculated and
used only if the elapsed time of the stopped calibration is longer
than the minimal calibration duration for reasonable precision;
otherwise, previous calibration data are used. The default calibration
duration is increased from 25 to 30 seconds for better precision.

The calibration data are stored without truncation. Before they are
used for time conversion, they are proportionally scaled down with
a factor, which is calculated based on the calibration duration and
the time to be converted, to maximize precision and to avoid possible
overflow in the 32-bit calculation.

Resolves: LXSWIREF-1747
Related: QTI9X07-2191
Signed-off-by: Dylan Zhang <dyzhang@sierrawireless.com>
Signed-off-by: Dragan Marinkovic <dmarinkovi@sierrawireless.com>
Modified-by: Dragan Marinkovic <dmarinkovi@sierrawireless.com>
(from 879ce3380aafdad02082c82b10114f52a7cdae70)
Change-Id: Ie39884c9117e3c5cde9321abd729305a96c4c9aa