ehci-msm-hsic: Do explicit suspend of bus
Instead of relying on runtime suspend framework to suspend HSIC
peripherals and bus, perform an explicit suspend of this bus when
the system goes to sleep.

Enable remote power-up of SMSC95xx HSIC slave to avoid powering
down the PHY and hence re-acquiring IP address on power-up. Remove
EMAC driver as RGMII pins are not accessible on WP76.

This is a re-implementation of a similar commit in 3.18 kernel:
    40abeeb HSIC: suspend related fixes

Resolves: LXSWIREF-1787
Change-Id: I87bcc00e310ea52a520cfb0693f93f93b2d7ab50
Signed-off-by: Zoran Markovic <zmarkovic@sierrawireless.com>