msm_serial: Find console on ports other than ttyMSM0
Function msm_console_setup() assumed that the console is always
assigned to the first line. This caused console never to be setup
in some cases when multiple consoles are listed on command line.

Traverse the array of available UARTs to find the first mapped
one. Assign console to this uart. This implementation is similar
to uart_get_console() that some other drivers are using.

Resolves: LXSWIREF-1815
Change-Id: Ibe90be10c33bb32959e0f151d9e28985fc614029
Relates: QTI9X07-5067 LXSWIREF-1821
Signed-off-by: Zoran Markovic <zmarkovic@sierrawireless.com>