audio: Integrate external Wolfson codec with MSM audio
Build Wolfson/Cirrys Logic SoC driver as kernel module, and
move its sources out of the kernel tree. The reason being is
that QCOM SoC kernel audio support is not in the kernel tree
any more.

Multi Function Device (MFD) part of WM8944 is still in the
kernel tree, as it must be available at the time of the SoC
kernel module loading.

Resolves: LXSWIREF-1778
Change-Id: Ic51391d6b8f1b5f5b68435ebeda8da0d340b1275
Signed-off-by: Dragan Marinkovic <dmarinkovi@sierrawireless.com>