kbuild: Add PYTHON3 variable
Yocto 3.1 includes only "python3" in hosttools and not "python" or
"python2". Add PYTHON3 variable, as done in the linux-yocto Makefile,
and use it to launch the python scripts for Kernel build. Retaining
PYTHON variable, since it is kept in linux-yocto as well as mainline
kernel Makefile.

Reference commits in linux-yocto:
 e9781b52d4e0e3381351d3483cfd173a968dcbe6
 94f7345b712405b79647a6a4bf8ccbd0d78fa69d

Resolves: LXSWIREF-1866
Change-Id: Ic757bf3057b85e8edb8b56218e1196cc400002ff
Signed-off-by: Steve MacDonald <smacdonald@sierrawireless.com>
Modified-by: Anju Kunjumon <akunjumon@sierrawireless.com>