uart: Support HS and LS drivers
Enable UART1 and UART2 to be able to operate with either high-speed or
low-speed drivers, based on the service that is mapped against the UART.

Two configurations are kept in the device tree for each UART. One with
"msm_serial" driver, to support low-speed services such as console, DM,
etc., and another with "msm_serial_hs" driver, to support high-speed
services such as AT, NMEA, etc. Each supported service is
programmatically mapped to a unique speed by sierra_serial.c. Each
driver probe() will check these mappings and fail, if a match is not
found.

- uart_create_sysfs_config() function returns -1 when a good mapping is
  not found for the driver against the UART.
- Do not crash on errors. Print and return error, to facilitate easier
  debugging.

Change in behaviour compared to mdm9x28le20-ref:
- UART2 HS mappings from mdm9x28le20-swi-wp76xx commits 61e515e and
  37bd0ae are brought in.
- A check is added in sierra_serial.c to catch programming error that
  maps two speeds to the same service.
- Invalid mappings are treated as BS_UART_FUNC_INVALID, by both HS and
  LS drivers.
- Only the successful driver will create a "config" entry under sysfs.
- The debug prints contain not the tty name, but the device name. The
  tty associated with the device can be found from the accompanying
  logs.
- device probe returns -ENODEV instead of -EPERM, in case of mapping
  failure.

Resolves: LXSWIREF-1894
Change-Id: I863c293d3fff2205cfc21e771c6d5777068b7ac4
Signed-off-by: Anju Kunjumon <akunjumon@sierrawireless.com>