tty: Fix basic issues in msm_serial_hs
There are a few issues found around Rx wakeup mechanism and msm_hs_isr
of msm_serial_hs.c.

1. Rx cannot wakeup the device: A wake-up mechanism is missing in the
msm_hs_wakeup_isr. When the device enters run-time suspend mode, the
clocks are turned off. The first Rx byte is supposed to be able to
wakeup the device; however, this part to turn on the clocks is missing
in the wakeup isr. A call to msm_hs_resource_vote() is added to
facilitate the same. msm_hs_resource_unvote() follow-up call ensures
that the device doesn't stay awake indefinitely. It will enter suspend
mode if there is no further Rx or Tx activity on the lines, within
"autosuspend_delay" time.
Note: The fix is slightly different from 55b034d702a9 in 3.18
mdm9x28le20-swi-wp76xx. The fix there could cause the device to stay
awake indefinitely if there is no further Rx or Tx activity on the
lines (as observed in cdb000a61c3b).

2. Loss of first byte: This is an expected behaviour associated with
the run-time suspend feature of the driver. When the device is in
suspend mode, the first received byte is lost in waking up the device.
The user experience improvement of fb534411720f from 3.18
mdm9x28le20-swi-wp76xx is brought in. "To improve the user experience,
we increase this suspend timer to 5 seconds from 200 ms. Please note,
this suspend time is not very accurate because it relies on a software
timer."

3. Loss of first two bytes: An "ignore" flag exists in the driver, to
prevent the device from being woken up by a stray interrupt. The wakeup
interrupt, which is essentially "a negative-edge trigger" configuration
of UART Rx line, is disabled by the wakeup_isr when resuming the device
from suspend. When there is an Rx activity afterwards, with the
interrupt disabled, the kernel holds the interrupt pending and will try
to service it on the next enable. The enable happens in the suspend
routine. In effect, whenever there is an Rx activity (aside from the
first wake-up Rx byte), there will be a stray Rx wake-up interrupt on
the next suspend. The flag is supposed to make the wake-up ISR ignore
this stray interrupt.
However, the driver sets the flag at both start-up and suspend,
irresepective of the Rx activity that is the real source of the stray
interrupt. So when there is a suspend without a preceding Rx activity,
there will not be any stray interrupt, yet the ignore flag will be set.
A genuine Rx wake-up afterwards would be ignored. It will take one more
byte to really wakeup the system. In total two bytes are lost. First one
is ignored, the second one is used for wakeup. This can be prevented by
setting the "ignore" flag only on Rx activity, which is the source of
the stray interrupt.
Note: This solution is different from "b5fded060761" that addresses the
same issue in 3.18 mdm9x28le20-swi-wp76xx branch.

4. Bug warning in CTS interrupt handler.
CTS interrupt handler uses msm_hs_resource_vote to turn on the clock
before clearing its interrupt flag, but this function is not reentrant.
Guard the non-reentrant part of the function to let it be used from
both wakeup isr and CTS interrupt handler context.
Note: This solution is provided by Qualcomm and is different from
"518ba655c50a" that addresses the same issue in
3.18 mdm9x28le20-swi-wp76xx branch.

Resolves: LXSWIREF-1894
Change-Id: I1602ada696bc34918bed125d1a6c3b51f6697bbc
Signed-off-by: Anju Kunjumon <akunjumon@sierrawireless.com>