build: Resolve TypeError reported in gcc-wrapper.py
In python3, the stderr of subprocess returns a binary stream
(b'some error') unless specified to use text mode. The script is
designed to expect a string and uses some string operations on this
returned value resulting in the following TypeError.

Traceback (most recent call last):
  File "scripts/gcc-wrapper.py", line 103, in <module>
    status = run_gcc()
  File "scripts/gcc-wrapper.py", line 89, in run_gcc
    interpret_warning(line)
  File "scripts/gcc-wrapper.py", line 60, in interpret_warning
    line = line.rstrip('\n')
TypeError: a bytes-like object is required, not 'str'

The solution is to specify the use of text mode, using
"universal_newlines=True" parameter, as it returns a string regardless
of Python version. The better named alias "text" will work only on
python >= 3.7, and hence not chosen.

Resolves: LXSWIREF-1927
Change-Id: I866d5a5c206427b2772332617357d4cad7438ab9
Signed-off-by: Anju Kunjumon <akunjumon@sierrawireless.com>